move AsyncWriter to separate file
