+++ src/reduct/storage/async_writer.h
@@ -1,0 +1,23 @@
+// Copyright 2022 Alexey Timin
+
+#ifndef REDUCT_STORAGE_ASYNC_WRITER_H
+#define REDUCT_STORAGE_ASYNC_WRITER_H
+
+#include <filesystem>
+
+#include "reduct/async/io.h"
+#include "reduct/proto/storage/entry.pb.h"
+
+namespace reduct::storage {
+
+struct AsyncWriterParameters {
+  std::filesystem::path path;
+  int record_index;
+  size_t size;
+};
+
+async::IAsyncWriter::UPtr BuildAsyncWriter(const proto::Block& block, AsyncWriterParameters parameters);
+
+}  // namespace reduct::storage
+
+#endif  // REDUCT_STORAGE_ASYNC_WRITER_H
