+++ src/reduct/storage/entry.cc
@@ -13,6 +13,7 @@
 #include "reduct/core/logger.h"
 #include "reduct/core/result.h"
 #include "reduct/proto/storage/entry.pb.h"
+#include "reduct/storage/async_writer.h"
 #include "reduct/storage/block_helpers.h"
 
 namespace reduct::storage {
@@ -26,70 +27,6 @@ auto to_time_t = IEntry::Time::clock::to_time_t;
 
 namespace fs = std::filesystem;
 
-/**
- * @class Asynchronous writer
- * @brief Writes chunks of data into pre-allocated block
- */
-class AsyncWriter : public async::IAsyncWriter {
- public:
-  struct Parameters {
-    fs::path path;
-    int record_index;
-    size_t size;
-  };
-
-  AsyncWriter(const proto::Block& block, Parameters parameters)
-      : ts_(block.begin_time()), parameters_(parameters), writen_size_{} {
-    file_ = std::ofstream(parameters_.path, std::ios::out | std::ios::in | std::ios::binary);
-    file_.seekp(block.records(parameters_.record_index).begin());
-  }
-
-  Error Write(std::string_view chunk, bool last) noexcept override {
-    if (!file_) {
-      UpdateRecord(proto::Record::kErrored);
-      return {.code = 500, .message = "Bad file"};
-    }
-
-    writen_size_ += chunk.size();
-    if (writen_size_ > parameters_.size) {
-      UpdateRecord(proto::Record::kErrored);
-      return {.code = 413, .message = "Content is bigger than in content-length"};
-    }
-
-    if (!file_.write(chunk.data(), chunk.size())) {
-      UpdateRecord(proto::Record::kErrored);
-      return {.code = 500, .message = "Failed to write a chunk into a block"};
-    }
-
-    if (last) {
-      UpdateRecord(proto::Record::kFinished);
-      file_ << std::flush;
-    }
-
-    return Error::kOk;
-  }
-
- private:
-  void UpdateRecord(proto::Record::State state) {
-    proto::Block block;
-    if (auto err = LoadBlockByTimestamp(parameters_.path.parent_path(), ts_, &block)) {
-      LOG_ERROR("{}", err);
-      return;
-    }
-
-    block.mutable_records(parameters_.record_index)->set_state(state);
-    if (auto err = SaveBlock(parameters_.path.parent_path(), block)) {
-      LOG_ERROR("{}", err);
-    }
-  }
-
-  std::ofstream file_;
-  Timestamp ts_;
-  Parameters parameters_;
-  size_t writen_size_;
-};
-
-Error SaveBlock(proto::Block block);
 class Entry : public IEntry {
  public:
   /**
@@ -208,12 +145,8 @@ class Entry : public IEntry {
     }
 
     return {
-        std::make_unique<AsyncWriter>(block,
-                                      AsyncWriter::Parameters{
-                                          .path = BlockPath(full_path_, block),
-                                          .record_index = block.records_size() - 1,
-                                          .size = size,
-                                      }),
+        BuildAsyncWriter(
+            block, {.path = BlockPath(full_path_, block), .record_index = block.records_size() - 1, .size = size}),
         SaveBlock(block),
     };
   }
